requirements.txt
ultralytics==8.3.117
python-dotenv==1.1.0
backports.tarfile==1.2.0
importlib-metadata==8.0.0
ipykernel==6.29.5
jaraco.collections==5.1.0
pip-chill==1.0.3
tomli==2.0.1
